+++ PureEdgeSim/com/mechalikh/pureedgesim/TasksGenerator/DefaultTasksGenerator.java
@@ -19,8 +19,7 @@ public class DefaultTasksGenerator extends TasksGenerator {
 		double simulationTime = (SimulationParameters.SIMULATION_TIME - SimulationParameters.INITIALIZATION_TIME) / 60;
 		for (int dev = 0; dev < datacentersList.size(); dev++) { // for each device
 			if (datacentersList.get(dev).isGeneratingTasks()) {
-				int app = new Random().nextInt(SimulationParameters.APPS_COUNT); // pickup a random application type for
-																					// every device
+				int app = new Random().nextInt(SimulationParameters.APPLICATIONS_LIST.size()); // pickup a random application type for every device
 				datacentersList.get(dev).setApplication(app); // assign this application to that device
 				for (int st = 0; st < simulationTime; st++) { // for each minute
 					// generating tasks
@@ -30,7 +29,7 @@ public class DefaultTasksGenerator extends TasksGenerator {
 					// Then pick up random second in this minute "st"
 					// Shift the time by the defined value "INITIALIZATION_TIME"
 					// in order to start after generating all the resources
-					time += new Random().nextInt(59) + SimulationParameters.INITIALIZATION_TIME;
+					time += new Random().nextInt(15)+ SimulationParameters.INITIALIZATION_TIME;
 					insert(time, app, dev);
 				}
 			}
@@ -39,22 +38,23 @@ public class DefaultTasksGenerator extends TasksGenerator {
 	}
 
 	private void insert(int time, int app, int dev) {
-		double maxLatency = (long) SimulationParameters.APPLICATIONS_TABLE[app][0]; // Load length from application file
-		long length = (long) SimulationParameters.APPLICATIONS_TABLE[app][3]; // Load length from application file
-		long requestSize = (long) SimulationParameters.APPLICATIONS_TABLE[app][1];
-		long outputSize = (long) SimulationParameters.APPLICATIONS_TABLE[app][2];
-		int pesNumber = (int) SimulationParameters.APPLICATIONS_TABLE[app][4];
-		long containerSize = (int) SimulationParameters.APPLICATIONS_TABLE[app][5]; // the size of the container
-		Task[] task = new Task[SimulationParameters.TASKS_PER_EDGE_DEVICE_PER_MINUTES];
+		double maxLatency = SimulationParameters.APPLICATIONS_LIST.get(app).getLatency(); // Load length from application file
+		long length = (long) SimulationParameters.APPLICATIONS_LIST.get(app).getTaskLength(); // Load length from application file
+		long requestSize = SimulationParameters.APPLICATIONS_LIST.get(app).getRequestSize();
+		long outputSize = SimulationParameters.APPLICATIONS_LIST.get(app).getResultsSize();
+		int pesNumber = SimulationParameters.APPLICATIONS_LIST.get(app).getNumberOfCores();
+		long containerSize = SimulationParameters.APPLICATIONS_LIST.get(app).getContainerSize(); // the size of the container
+		Task[] task = new Task[SimulationParameters.APPLICATIONS_LIST.get(app).getRate()];
 		int id;
 
 		// generate tasks for every edge device
-		for (int i = 0; i < SimulationParameters.TASKS_PER_EDGE_DEVICE_PER_MINUTES; i++) {
+		for (int i = 0; i < SimulationParameters.APPLICATIONS_LIST.get(app).getRate(); i++) {
 			id = taskList.size();
 			UtilizationModel utilizationModeldynamic = new UtilizationModelDynamic();
 			task[i] = new Task(id, length, pesNumber);
 			task[i].setFileSize(requestSize).setOutputSize(outputSize).setUtilizationModelBw(utilizationModeldynamic)
 					.setUtilizationModelRam(utilizationModeldynamic).setUtilizationModelCpu(new UtilizationModelFull());
+			time+=60/SimulationParameters.APPLICATIONS_LIST.get(app).getRate();
 			task[i].setTime(time);
 			task[i].setContainerSize(containerSize);
 			task[i].setMaxLatency(maxLatency);