+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 # Changelog
+## Version 3.0 (mai 5th 2020)
+
+*   Added MinTimeBetweenEvents to MainApplication.java .
+
+*   Fixed bugs (network model, charts, output).
+
+*   Improved tasks generator.
+
+*   New example "Example8" with a fuzzy logic based workload orchestration.
 
 ## Version 2.4 (apr 7th 2020)